Added more details to MatchDB, decoupled PlayerStats from Match data

+++ app/src/main/java/com/saladmuffin/leagueapi/databases/ChampionDB.java
@@ -29,7 +29,7 @@ public class ChampionDB {
     public static final String SQL_DELETE_ENTRIES =
             "DROP TABLE IF EXISTS " + ChampionEntry.TABLE_NAME;
 
-    public static String queryChampionById(int id) {
+    public static String queryChampionById(long id) {
         return "SELECT * FROM " + ChampionEntry.TABLE_NAME + " WHERE "
                 + ChampionEntry.COLUMN_NAME_CHAMPION_ID + "=" + id;
     }
